Adds a tiny Snyk vuln integration
This mimics the github code fairly closely: has a (working link) placeholder to check for vulns for the specific package, then if JS is available, it'll make an XHR request to Snyk's API and update with any vulns in the package.

The link has a number of states:

- [Test for vulnerabilities] on Snyk (non JS state)
- [No known vulnerabilities] found on Snyk
- [1 vulnerable dependency] found on Snyk (package has vulns in dependencies)
- [4 vulnerable dependencies] found on Snyk
- [1 known vulnerability] found on Snyk (package *is* vulnerable)
- [5 known vulnerabilities] found on Snyk

This change also updates the csp to include snyk.io's specific npm API endpoint in the `connectSrc` list.